+++ client/src/components/pages/FormPages.js
@@ -67,6 +67,8 @@ import {
   invalidSecurityGroupRuleText,
   storageChangeDisabledCallback,
   vpnServersHelperText,
+  powerImageFetch,
+  powerStoragePoolFetch,
 } from "../../lib/forms";
 import {
   invalidCidr,
@@ -140,18 +142,26 @@ const AppIdPage = (craig) => {
     <AppIdTemplate
       docs={RenderDocs("appid")}
       appid={craig.store.json.appid}
-      disableSave={disableSave}
+      disableSave={function (field, stateData, componentProps) {
+        // field is passed here but the goal is to be able to ignore the
+        // parameter entirely
+        return craig.appid.shouldDisableSave(stateData, componentProps);
+      }}
+      forceOpen={function (stateData, componentProps) {
+        // goal here is to be able to pass one param `shouldDisableSave`
+        // and handle both disable and force Open
+        return craig.appid.shouldDisableSave(stateData, componentProps);
+      }}
       onDelete={craig.appid.delete}
       onSave={craig.appid.save}
       onSubmit={craig.appid.create}
       propsMatchState={propsMatchState}
-      forceOpen={forceShowForm}
       craig={craig}
       resourceGroups={splat(craig.store.json.resource_groups, "name")}
-      invalidCallback={invalidName("appid")}
-      invalidTextCallback={invalidNameText("appid")}
-      invalidKeyCallback={invalidName("appid_key")}
-      invalidKeyTextCallback={invalidNameText("appid_key")}
+      invalidCallback={craig.appid.name.invalid}
+      invalidTextCallback={craig.appid.name.invalidText}
+      invalidKeyCallback={craig.appid.keys.name.invalid}
+      invalidKeyTextCallback={craig.appid.keys.name.invalidText}
       onKeySave={craig.appid.keys.save}
       onKeyDelete={craig.appid.keys.delete}
       onKeySubmit={craig.appid.keys.create}
@@ -418,6 +428,9 @@ const KeyManagementPage = (craig) => {
           craig.store.json._options.fs_cloud
         );
       }}
+      selectEndpoint={
+        craig.store.json._options.endpoints === "public-and-private"
+      }
       resourceGroups={splat(craig.store.json.resource_groups, "name")}
       invalidCallback={invalidName("key_management")}
       invalidTextCallback={invalidNameText("key_management")}
@@ -493,17 +506,17 @@ const ObjectStoragePage = (craig) => {
       onSubmit={craig.object_storage.create}
       propsMatchState={propsMatchState}
       forceOpen={forceShowForm}
-      craig={craig}
+      kmsList={splat(craig.store.json.key_management, "name")}
       resourceGroups={splat(craig.store.json.resource_groups, "name")}
+      craig={craig}
       cosPlans={cosPlans}
       encryptionKeys={craig.store.encryptionKeys}
-      kmsList={splat(craig.store.json.key_management, "name")}
-      invalidCallback={invalidName("object_storage")}
-      invalidTextCallback={invalidNameText("object_storage")}
-      invalidKeyCallback={invalidName("cos_keys")}
-      invalidKeyTextCallback={invalidNameText("cos_keys")}
-      invalidBucketCallback={invalidName("buckets")}
-      invalidBucketTextCallback={invalidNameText("buckets")}
+      invalidCallback={craig.object_storage.name.invalid}
+      invalidTextCallback={craig.object_storage.name.invalidText}
+      invalidKeyCallback={craig.object_storage.keys.name.invalid}
+      invalidKeyTextCallback={craig.object_storage.keys.name.invalidText}
+      invalidBucketCallback={craig.object_storage.buckets.name.invalid}
+      invalidBucketTextCallback={craig.object_storage.buckets.name.invalidText}
       onKeySave={craig.object_storage.keys.save}
       onKeyDelete={craig.object_storage.keys.delete}
       onKeySubmit={craig.object_storage.keys.create}
@@ -543,6 +556,12 @@ const NoPowerWorkspaceTile = () => {
 };
 
 const PowerInfraPage = (craig) => {
+  let powerImageMap = {};
+  craig.store.json._options.power_vs_zones.forEach((zone) => {
+    powerImageFetch(zone, fetch).then((zoneImages) => {
+      powerImageMap[zone] = zoneImages;
+    });
+  });
   return (
     <PowerVsWorkspacePage
       overrideTile={
@@ -622,7 +641,7 @@ const PowerInfraPage = (craig) => {
       invalidDnsCallbackText={() => {
         return "Invalid IP Address";
       }}
-      imageMap={require("../../lib/docs/power-image-map.json")}
+      imageMap={powerImageMap}
       onAttachmentSave={craig.power.attachments.save}
       disableAttachmentSave={storageChangeDisabledCallback}
     />
@@ -630,6 +649,12 @@ const PowerInfraPage = (craig) => {
 };
 
 const PowerVsInstances = (craig) => {
+  let powerStoragePoolMap = {};
+  craig.store.json._options.power_vs_zones.forEach((zone) => {
+    powerStoragePoolFetch(zone, fetch).then((zonePools) => {
+      powerStoragePoolMap[zone] = zonePools;
+    });
+  });
   return (
     <PowerVsInstancesPage
       overrideTile={
@@ -653,7 +678,7 @@ const PowerVsInstances = (craig) => {
         }
       })}
       power_instances={craig.store.json.power_instances}
-      storage_pool_map={powerStoragePoolRegionMap}
+      storage_pool_map={powerStoragePoolMap}
       power_volumes={craig.store.json.power_volumes}
       disableSave={disableSave}
       propsMatchState={propsMatchState}